Created card_sets.json, to start creating the card sets that the reward booster packs will come from.
Edited goal_objects to add author to books and get rid of a straggling artifact from a previous attempt.

Fixed add_books to users to add authors. Added a bit of test code to make sure that the dictionaries and methods were currently working. Will test more in-depth later. Also will add a testing-suite eventually, but for now I want to get things connected and the main methods working for Books, and then we can build a testing suite and finish the rest of the media objects.