Started the movie class
Need to figure out what to do about remakes by the same director. For example: The Ten Commandments:  Cecil B Demille could be counted for two different movies.

Maybe I will track year, as well, and if the years made are different then it will expand the title. Need to think on it a bit, if I even care.