Added the rest of the cards for Wildlife Mayhem to card_set_db.json!
Fixed set_pack_registry to make sure each card in each pack is unique (no duplicates per pack). The function looks ugly and is a bit longer than I'd like (~ 20 lines), so I may split it up or refactor it later, but I'll keep it for now and focus on building the app.

Lastly, fixed the Unique rarity checker, as that was still set for 18 rares instead of 12.