A lot of work today. In my fury, forgot to commit a couple of times.
Long story short, I connected cards.py and users.py so that I can add packs to the pack list, and it will automatically create the card sets along with the initalized user. There is also an update_card_sets button that will be utilized to refresh the card_sets as new sets are created.

Initial testing is promising, but there is still a long way to go on that front.