Got rid of the old card_sets.json file.
Replaced it with a new file, for the brand new Base Sets. I've decided to make a 400 card Equipment+ Base set. I'm excited about these new changes.

Added 41 new Commons to the set! 159 more to go!

Did a bit of manual testing for the expansion pack functions. They seem to be working well so far after a bit of bug fixing.

Cleaned up the code a bit, and fixed a small bug.