Created Events Class type, though still need to add methods.
Did some minor testing with adding tv_shows and seasons. There were some errors because I was trying to add a dictionary to it when it should be a list, instead. A dictionary in the long run is probably better, but it took me awhile to debug that, so I'll think on that more later. The only real benefit to a dictionary is if users add seasons out of order, which is definitely possible for reality-based tv shows, so it's likely worth doing the small amount of work to refactor it.

I want to refactor the entire goal_objects.py file for inheritance, now that I understand it a lot better. Will do that soon. I should probably do that before finishing the Events object, tbh, so that will likely be one of the next things I do.