Replaced all instances of completion_hours with completion_mins, since I've decided to use minutes to 1) avoid floats, and 2) be slightly more precise, and 3) more consistent between all the different goal object types.
Because of this, I might like to create a conversion function that will just take the user hour + min input and converts it strictly to minutes before calling these functions. It isn't the MAIN priority, but it may be easier for users to input in hours + minutes than just flat minutes.