Replaced users.py with users2.py!
Next I will be working on testing and making sure everything connects to one another correctly.

However, I'm going to take a short break (~ a week) off, and work on a much smaller project. I want to practice coming back to projects after a bit of time off, both to get a stronger psyche, but also to work on my comments and returning to code bases I may be a bit rusty on.